+++ src/lib.rs
@@ -106,8 +106,8 @@ pub use crate::traits::{
     EditContentRating, EditField, EditOriginalTitle, EditSortTitle, EditStudio, EditSummary,
     EditTagline, EditTags, EditTitle, EditYear, FieldValue, HasArtLock, HasArtUrl, HasCollections,
     HasCountries, HasDirectors, HasGenres, HasLabels, HasMoods, HasPosterLock, HasPosterUrl,
-    HasProducers, HasRoles, HasStyles, HasThemeLock, HasThemeUrl, HasWriters, PlayedUnplayed,
-    PlexObject, Ratable, Reload,
+    HasProducers, HasRoles, HasStyles, HasThemeLock, HasThemeUrl, HasWriters, Playable,
+    PlayedUnplayed, PlexObject, Ratable, Reload,
 };
 pub use crate::uri::PlexUri;
 pub use crate::util::{